Multi-plateforme (#35)
* Support for MacOS and Windows

* refactoring

* various bug fixes

* switched pipeline to CMake.

+++ libeasea/include/algorithms/moea/Casrea.h
@@ -103,7 +103,7 @@ typename Casrea<TIndividual, TRandom>::TPopulation Casrea<TIndividual, TRandom>:
 #ifdef USE_OPENMP
     EASEA_PRAGMA_OMP_PARALLEL
 #endif
-        for (size_t i = 0; i < offspring.size(); ++i)
+        for (int i = 0; i < offspring.size(); ++i)
         {
                 TI &child = offspring[i];
                 this->getMutation()(child);
@@ -140,11 +140,19 @@ void Casrea<TIndividual, TRandom>::makeOneGeneration(void)
 	size_t szPopDiv2 = (offspring.size() - TBaseArchive::m_archive.size()) / 2;
 	size_t icounter = TBaseArchive::m_archive.size();
 
+	//IMPORTANT: there is a problem below, I tried to fix it using the commented code, I reverted to the old version to bring back the old message. You may want to uncomment
+	//std::copy(TBaseArchive::m_archive.begin() /*+ std::max(0, static_cast<int>(TBaseArchive::m_archive.size()) - static_cast<int>(TBase::m_population.size()))*/, TBaseArchive::m_archive.begin() + std::min(TBaseArchive::m_archive.size(), TBase::m_population.size()), TBase::m_population.begin());
+	//if (TBaseArchive::m_archive.size() > TBase::m_population.size())
+	//	std::copy(TBaseArchive::m_archive.begin() + TBase::m_population.size(), TBaseArchive::m_archive.end() , std::back_inserter(TBase::m_population));
+
+	//LOG_MSG(msgType::DEBUG, "pop size: %d\narchive size: %d\n", TBase::m_population.size(), TBaseArchive::m_archive.size());
+
+	// IMPORTANT: old version below used totalSelection and was responsible for alignment errors and segfault, or just stop the program from working after N generations.
 	std::vector<TPtr> archPop;
 	for (size_t i = 0; i < TBaseArchive::m_archive.size(); ++i)
     		archPop.push_back(&TBaseArchive::m_archive[i]);
 	easea::operators::selection::totalSelection(archPop, TBase::m_population.begin(), TBase::m_population.end());
-	
+
 	for( size_t i = icounter; i < szPopDiv2; ++i)
 	{
 		std::uniform_int_distribution<size_t> dist(0, TBaseArchive::m_archive.size() - 1);
